refactor(LoadingSpinner): 로딩 스피너를 별도의 컴포넌트로 분리
